backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio Analysis
librosa==0.10.1
essentia==2.1b6.dev1110
madmom==0.16.1
scipy==1.11.4
numpy==1.24.3
soundfile==0.12.1
audioread==3.0.1

# OpenAI for Whisper API
openai==1.3.7

# Google Drive Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
beanie==1.23.6  # ODM for MongoDB

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
mutagen==1.47.0  # Audio metadata extraction

# CORS
fastapi-cors==0.0.6
